refactor message queue; enforce check for faction and friendlies while caching results; rename appropriately

+++ ImprovedAI/Data/Scripts/ImprovedAI/IAIMessage.cs
@@ -5,7 +5,7 @@ using VRageMath;
 
 namespace ImprovedAI.Messages
 {
-    public enum MessageTopics : ushort
+    public enum MessageTopics : byte
     {
         DRONE_REGISTRATION,
         DRONE_REPORTS,
